requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
amqp
asgiref
async-timeout
attrs
billiard
boto3
botocore
celery
certifi
cffi
channels
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cloudinary
colorama
cryptography
Deprecated
diff-match-patch
distlib
dj-database-url
Django==6.0
django-anymail
django-auditlog
django-celery-beat
django-cloudinary-storage
django-cors-headers
django-encrypted-model-fields
django-environ
django-filter
django-import-export
django-jazzmin
django-phone-verify
django-phonenumber-field
django-phonenumber-field[phonenumbers]
django-ratelimit
django-redis
django-storages
django-zoho-zeptomail
djangorestframework
djangorestframework-simplejwt
drf-api-logger
drf-spectacular
drf-yasg
environs
filelock
flower
frozenlist
grep uuid
google-auth
gunicorn
hiredis
humanize
idna
inflection
jmespath
jsonschema
jsonschema-specifications
kombu
marshmallow
multidict
nexmo
packaging
phonenumbers
pillow
pipenv
platformdirs
prometheus_client
prompt_toolkit
propcache
psycopg2
psycopg2-binary
pycparser
pycryptodome
pydantic
PyJWT
python-dateutil
python-decouple
python-dotenv
pytz
PyYAML
rave-python
redis
referencing
requests
rpds-py
s3transfer
setuptools
shortuuid
six
sqlparse
stripe
tablib
tornado
twilio
typing_extensions
tzdata
uritemplate
urllib3
uuid6
vine
virtualenv
wcwidth
wheel
whitenoise
wrapt
yarl